Add BSB Tx Enable Control
Move the UEXT RX on GPIO35 (input only), and use GPIO32 instead for controlling whether the BSB-TXD pin can drive the optocoupler. This basically replaces the +3.3V connection on the emitter of the invert transistor.

What this means is that when the ESP32 has not yet booted, or if it crashes along the way, the BSB bus is not being driven into TX mode by the pull-down bootstrap resistor. This should stop BSB bus short detection errors when the unit is powered off or has not yet started up.

This will require custom code in the BSB firmware setup (BSB_LAN_custom_setup.h) to configure the enable pin to be an output and driven high. Until that is done the BSB bus remains passive.