Change BSB-to-UEXT Tx Pin Assignment
Follow the BSB-LAN repo in moving TX onto the UEXT-SSEL pin instead of the UEXT-TX pin. This is done to avoid issues with the TX pin needing to be pulled low during the boot phase, but that in turn would "short" out the BSB bus until everything is booted up.